+++ conda_env/installers/conda.py
@@ -8,7 +8,7 @@ from conda.core.solve import Solver
 from conda.models.channel import Channel, prioritize_channels
 
 
-def install(prefix, specs, args, env, prune=False):
+def install(prefix, specs, args, env, *_, **kwargs):
     # TODO: support all various ways this happens
     # Including 'nodefaults' in the channels list disables the defaults
     new_specs = []
@@ -33,7 +33,7 @@ def install(prefix, specs, args, env, prune=False):
     subdirs = IndexedSet(basename(url) for url in _channel_priority_map)
 
     solver = Solver(prefix, channels, subdirs, specs_to_add=specs)
-    unlink_link_transaction = solver.solve_for_transaction(prune=prune)
+    unlink_link_transaction = solver.solve_for_transaction(prune=getattr(args, 'prune', False))
 
     pfe = unlink_link_transaction.get_pfe()
     pfe.execute()