+++ tests/test_create.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from glob import glob
 import json
 from json import loads as json_loads
-from logging import DEBUG, getLogger
+from logging import DEBUG, INFO, getLogger
 import os
 from os.path import basename, dirname, exists, isdir, isfile, join, lexists, relpath
 from random import sample
@@ -26,7 +26,7 @@ import requests
 
 from conda import CondaError, CondaMultiError
 from conda._vendor.auxlib.entity import EntityEncoder
-from conda.base.constants import CONDA_TARBALL_EXTENSION, PACKAGE_CACHE_MAGIC_FILE
+from conda.base.constants import CONDA_TARBALL_EXTENSION, PACKAGE_CACHE_MAGIC_FILE, SafetyChecks
 from conda.base.context import Context, context, reset_context
 from conda.cli.main import generate_parser, init_loggers
 from conda.common.compat import PY2, iteritems, itervalues, text_type
@@ -57,8 +57,10 @@ except ImportError:
 
 
 log = getLogger(__name__)
-TRACE, DEBUG = TRACE, DEBUG  # these are so the imports aren't cleared, but it's easy to switch back and forth
+TRACE, DEBUG, INFO = TRACE, DEBUG, INFO  # these are so the imports aren't cleared, but it's easy to switch back and forth
 TEST_LOG_LEVEL = DEBUG
+stderr_log_level(TEST_LOG_LEVEL, 'conda')
+stderr_log_level(TEST_LOG_LEVEL, 'requests')
 PYTHON_BINARY = 'python.exe' if on_win else 'bin/python'
 BIN_DIRECTORY = 'Scripts' if on_win else 'bin'
 UINCODE_CHARACTERS = u"ōγђ家固한"
@@ -312,33 +314,58 @@ class IntegrationTests(TestCase):
             assert not package_is_installed(prefix, 'flask')
             assert_package_is_installed(prefix, 'python-3')
 
-    def test_skip_safety_checks_basic(self):
+    def test_safety_checks(self):
+        # This test uses https://anaconda.org/conda-test/spiffy-test-app/0.5/download/noarch/spiffy-test-app-0.5-pyh6afbcc8_0.tar.bz2
+        # which is a modification of https://anaconda.org/conda-test/spiffy-test-app/1.0/download/noarch/spiffy-test-app-1.0-pyh6afabb7_0.tar.bz2
+        # as documented in info/README within that package.
+        # I also had to fix the post-link script in the package by adding quotation marks to handle
+        # spaces in path names.
+
         with make_temp_env() as prefix:
             with open(join(prefix, 'condarc'), 'a') as fh:
-                fh.write("skip_safety_checks: true\n")
+                fh.write("safety_checks: enabled\n")
             reload_config(prefix)
-            assert context.skip_safety_checks is True
+            assert context.safety_checks is SafetyChecks.enabled
 
-            run_command(Commands.INSTALL, prefix, 'python=3.5')
-            assert exists(join(prefix, PYTHON_BINARY))
-            assert_package_is_installed(prefix, 'python-3')
+            with pytest.raises(CondaMultiError) as exc:
+                run_command(Commands.INSTALL, prefix, '-c conda-test spiffy-test-app=0.5')
 
-            run_command(Commands.INSTALL, prefix, 'flask=0.10')
-            assert_package_is_installed(prefix, 'flask-0.10.1')
-            assert_package_is_installed(prefix, 'python-3')
+            error_message = text_type(exc.value)
+            message1 = dals("""
+            The path 'site-packages/spiffy_test_app-1.0-py2.7.egg-info/top_level.txt'
+            has an incorrect size.
+              reported size: 32 bytes
+              actual size: 16 bytes
+            """)
+            message2 = dals("""
+            The path 'site-packages/spiffy_test_app/__init__.py'
+            has a sha256 mismatch.
+              reported sha256: 1234567890123456789012345678901234567890123456789012345678901234
+              actual sha256: 32d822669b582f82da97225f69e3ef01ab8b63094e447a9acca148a6e79afbed
+            """)
+            assert message1 in error_message
+            assert message2 in error_message
 
-            run_command(Commands.INSTALL, prefix, '--force', 'flask=0.10')
-            assert_package_is_installed(prefix, 'flask-0.10.1')
-            assert_package_is_installed(prefix, 'python-3')
+            with open(join(prefix, 'condarc'), 'a') as fh:
+                fh.write("safety_checks: warn\n")
+            reload_config(prefix)
+            assert context.safety_checks is SafetyChecks.warn
 
-            run_command(Commands.UPDATE, prefix, 'flask')
-            assert not package_is_installed(prefix, 'flask-0.10.1')
-            assert_package_is_installed(prefix, 'flask')
-            assert_package_is_installed(prefix, 'python-3')
+            stdout, stderr = run_command(Commands.INSTALL, prefix, '-c conda-test spiffy-test-app=0.5')
+            assert message1 in stderr
+            assert message2 in stderr
+            assert package_is_installed(prefix, "spiffy-test-app-0.5")
 
-            run_command(Commands.REMOVE, prefix, 'flask')
-            assert not package_is_installed(prefix, 'flask-0.')
-            assert_package_is_installed(prefix, 'python-3')
+        with make_temp_env() as prefix:
+            with open(join(prefix, 'condarc'), 'a') as fh:
+                fh.write("safety_checks: disabled\n")
+            reload_config(prefix)
+            assert context.safety_checks is SafetyChecks.disabled
+
+            stdout, stderr = run_command(Commands.INSTALL, prefix, '-c conda-test spiffy-test-app=0.5')
+            assert message1 not in stderr
+            assert message2 not in stderr
+            assert package_is_installed(prefix, "spiffy-test-app-0.5")
 
     @pytest.mark.xfail(strict=True)
     def test_non_root_conda(self):
@@ -372,9 +399,22 @@ class IntegrationTests(TestCase):
         try:
             prefix = make_temp_prefix(str(uuid4())[:7])
 
+            stdout, stderr = run_command(Commands.CREATE, prefix, "python=3.5 --json --dry-run", use_exception_handler=True)
+            assert_json_parsable(stdout)
+            assert not stderr
+
+            # regression test for #5825
+            # contents of LINK and UNLINK is expected to have Dist format
+            json_obj = json.loads(stdout)
+            dist_dump = json_obj['actions']['LINK'][0]
+            assert 'dist_name' in dist_dump
+
             stdout, stderr = run_command(Commands.CREATE, prefix, "python=3.5 --json")
             assert_json_parsable(stdout)
             assert not stderr
+            json_obj = json.loads(stdout)
+            dist_dump = json_obj['actions']['LINK'][0]
+            assert 'dist_name' in dist_dump
 
             stdout, stderr = run_command(Commands.INSTALL, prefix, 'flask=0.10 --json')
             assert_json_parsable(stdout)
@@ -401,6 +441,12 @@ class IntegrationTests(TestCase):
             assert not stderr
             assert not package_is_installed(prefix, 'flask-0.')
             assert_package_is_installed(prefix, 'python-3')
+
+            # regression test for #5825
+            # contents of LINK and UNLINK is expected to have Dist format
+            json_obj = json.loads(stdout)
+            dist_dump = json_obj['actions']['UNLINK'][0]
+            assert 'dist_name' in dist_dump
 
             stdout, stderr = run_command(Commands.LIST, prefix, '--revisions --json')
             assert not stderr
@@ -1035,8 +1081,9 @@ class IntegrationTests(TestCase):
         stdout, stderr = run_command(Commands.CREATE, prefix, "flask", "--dry-run", "--json", use_exception_handler=True)
 
         loaded = json.loads(stdout)
-        assert "python" in "\n".join(loaded['actions']['LINK'])
-        assert "flask" in "\n".join(loaded['actions']['LINK'])
+        names = set(d['name'] for d in loaded['actions']['LINK'])
+        assert "python" in names
+        assert "flask" in names
 
     def test_packages_not_found(self):
         with make_temp_env() as prefix: